refactor: 게이미피케이션 상수 패키지 구조 개선
