+++ src/main/java/com/eventitta/gamification/constants/GamificationErrorCodes.java
@@ -1,4 +1,4 @@
-package com.eventitta.gamification.constant;
+package com.eventitta.gamification.constants;
 
 public final class GamificationErrorCodes {
